abstracted function and added store
